requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
scikit-surprise==1.1.4
PyMySQL==1.1.1
requests==2.32.3
beautifulsoup4==4.12.3
